Full script to generate sample messages on the bus

+++ script/test.rb
@@ -4,12 +4,23 @@ require 'optparse'
 require 'lims-management-app'
 require 'json'
 
+# The following script is used to generate messages 
+# on the bus for dev/test purpose for the lims-bridge-app.
+
 options = {}
 OptionParser.new do |opt|
-  opt.on('-q', '--quantity QUANTITY') do |qty|
-    options[:quantity] = qty
+  opt.on('-q', '--quantity QUANTITY') { |qty| options[:quantity] = qty.to_i }
+  opt.on('-c', '--create') { |o| options[:create] = true }
+  opt.on('-u', '--update') do |o| 
+    options[:create] = true
+    options[:update] = true
+  end
+  opt.on('-d', '--delete') do |o| 
+    options[:create] = true
+    options[:delete] = true
   end
 end.parse!
+options[:create] = true
 
 # helper
 def update_parameters(parameters)
@@ -61,31 +72,45 @@ if options[:quantity] == 1 || options[:quantity].nil?
         :concentration => 120,
         :sample_purified => false,
         :concentration_determined_by_which_method => "method"
+      },
+      :cellular_material => {
+        :lysed => false
+      },
+      :genotyping => {
+        :country_of_origin => "england",
+        :geographical_region => "europe",
+        :ethnicity => "english"
       }
     }
   }
 
-  response = RestClient.post("http://localhost:9292/samples",
-                             parameters.to_json,
-                             {'Content-Type' => 'application/json', 'Accept' => 'application/json'})
-  result = JSON.parse(response)
-  sample_uuid = result["sample"]["uuid"]
-  puts response
-  puts
+  if options[:create]
+    response = RestClient.post("http://localhost:9292/samples",
+                               parameters.to_json,
+                               {'Content-Type' => 'application/json', 'Accept' => 'application/json'})
+    result = JSON.parse(response)
+    sample_uuid = result["sample"]["uuid"]
+    puts response
+    puts
+  end
 
   # Update a sample
-  updated_parameters = update_parameters(parameters) 
-  response = RestClient.put("http://localhost:9292/#{sample_uuid}",
-                            updated_parameters[:sample].to_json,
-                            {'Content-Type' => 'application/json', 'Accept' => 'application/json'})
-  puts response
-  puts
+  if options[:update]
+    updated_parameters = update_parameters(parameters) 
+    response = RestClient.put("http://localhost:9292/#{sample_uuid}",
+                              updated_parameters[:sample].to_json,
+                              {'Content-Type' => 'application/json', 'Accept' => 'application/json'})
+    puts response
+    puts
+  end
 
   # Delete a sample
-  response = RestClient.delete("http://localhost:9292/#{sample_uuid}",
-                               {'Content-Type' => 'application/json', 'Accept' => 'application/json'})
-  puts response
-  puts
+  if options[:delete]
+    response = RestClient.delete("http://localhost:9292/#{sample_uuid}",
+                                 {'Content-Type' => 'application/json', 'Accept' => 'application/json'})
+    puts response
+    puts
+  end
 
 else
 
@@ -119,12 +144,51 @@ else
         :concentration => 120,
         :sample_purified => false,
         :concentration_determined_by_which_method => "method"
+      },
+      :cellular_material => {
+        :lysed => false
+      },
+      :genotyping => {
+        :country_of_origin => "england",
+        :geographical_region => "europe",
+        :ethnicity => "english"
       }
     }
   }
 
-  response = RestClient.post("http://localhost:9292/actions/bulk_create_sample",
-                             parameters.to_json,
-                             {'Content-Type' => 'application/json', 'Accept' => 'application/json'})
-  puts response
+  if options[:create]
+    response = RestClient.post("http://localhost:9292/actions/bulk_create_sample",
+                               parameters.to_json,
+                               {'Content-Type' => 'application/json', 'Accept' => 'application/json'})
+    result = JSON.parse(response)
+    sample_uuids = [].tap do |uuids|
+      result["bulk_create_sample"]["result"]["samples"].each do |sample|
+        uuids << sample["uuid"]
+      end
+    end
+    puts response
+    puts
+  end
+
+  # Sample bulk update
+  if options[:update]
+    updated_parameters = {:bulk_update_sample => update_parameters(parameters[:bulk_create_sample] - [:quantity]).merge({
+      "sample_uuids" => sample_uuids      
+    })}
+    response = RestClient.post("http://localhost:9292/actions/bulk_update_sample",
+                              updated_parameters.to_json,
+                              {'Content-Type' => 'application/json', 'Accept' => 'application/json'})
+    puts response
+    puts
+  end
+
+  # Bulk Delete samples
+  if options[:delete]
+    parameters = {:bulk_delete_sample => {:sample_uuids => sample_uuids}} 
+    response = RestClient.post("http://localhost:9292/actions/bulk_delete_sample",
+                               parameters.to_json,
+                                 {'Content-Type' => 'application/json', 'Accept' => 'application/json'})
+    puts response
+    puts
+  end
 end
